requirements.txt
discord.py
asyncio
python-dotenv
datetime
googletrans
pynacl